Fit mate scores in 16-bit for NNUE - test engine outputs random 16-bit scores between INT16_MIN and INT16_MAX. - scores within 1024 of each bounds are interpreted as mate(d) in X moves.
$ ./make.py -p test
[...]
File signatures:
% sha1sum stdout out1.pgn out2.pgn log training.csv
0bfd8677cca0fd88e86f5a0be2035b961536f576  stdout
5e6489fb1f3dbb489659511f9b0f1ddeb531ecf1  out1.pgn
b2a2471cf5937fa41aa38a02c5d247df57752558  out2.pgn
d0e99592f8488c517fc49db7e5a056cf5bc9434c  log
8fb744aa10c419994134ee51d08162313000d87a  training.csv

Overall signature:
% cat stdout out1.pgn out2.pgn log training.csv |sha1sum
04d1c94fc42f7b766c745b2bb8a1d5710467aae1  -

+++ src/engine.c
@@ -257,7 +257,7 @@ bool engine_bestmove(Worker *w, const Engine *e, int64_t *timeLeft, str_t *best,
                             info->score = atoi(token.buf);
                         else if (!strcmp(token.buf, "mate") && (tail = str_tok(tail, &token, " "))) {
                             const int movesToMate = atoi(token.buf);
-                            info->score = movesToMate < 0 ? INT_MIN - movesToMate : INT_MAX - movesToMate;
+                            info->score = movesToMate < 0 ? INT16_MIN - movesToMate : INT16_MAX - movesToMate;
                         } else
                             DIE("illegal syntax after 'score' in '%s'\n", line.buf);
                     }
